Checking that generation is OK

+++ generator/tests/glue.py
@@ -1,9 +1,13 @@
 #!/usr/bin/python3
 
-import random, unittest, os
+import random, unittest, os, shutil
 from glue import Glue
 
 class TestGlue(unittest.TestCase):
+    """
+    Parses a simple file containing a block and tests that its input, output and
+    parameter is recognized
+    """
     def test_glue_basic(self):
         glue = Glue()
         glue.parse(self.get_file('files/Gain.h'))
@@ -39,6 +43,10 @@ class TestGlue(unittest.TestCase):
         self.assertEqual('getOutput()', output.get())
         self.assertEqual('float', output.type)
 
+    """
+    Parses a simple file and tests that inputs, outputs and parameters are OK,
+    it contains some multiple I/Os
+    """
     def test_glue_multiple(self):
         glue = Glue()
         glue.parse(self.get_file('files/Gains.h'))
@@ -71,16 +79,49 @@ class TestGlue(unittest.TestCase):
         self.assertEqual('float', output.type)
         self.assertEqual('float', output.accessor_type())
 
+    """
+    Testing that files are indeed generated
+    """
+    def test_glue_generation_basic(self):
+        glue = Glue()
+        glue.parse(self.get_file('files/Gain.h'))
+        glue.parse(self.get_file('files/Gains.h'))
+        self.generate(glue)
+
+        self.assertTrue(os.path.isdir(self.get_output('glue')))
+        self.assertTrue(os.path.isfile(self.get_output('glue/glue.cpp')))
+        self.assertTrue(os.path.isfile(self.get_output('glue/convert.h')))
+        self.assertTrue(os.path.isfile(self.get_output('glue/deserialize.h')))
+        self.assertTrue(os.path.isfile(self.get_output('glue/GlueTypes.h')))
+
+        self.assertTrue(os.path.isfile(self.get_output('glue/GlueGain.cpp')))
+        self.assertTrue(os.path.isfile(self.get_output('glue/GlueGain.h')))
+        self.assertTrue(os.path.isfile(self.get_output('glue/GlueGains.cpp')))
+        self.assertTrue(os.path.isfile(self.get_output('glue/GlueGains.h')))
+
+        self.assertTrue(os.path.isdir(self.get_output('web')))
+        self.assertTrue(os.path.isdir(self.get_output('web/blocks.js')))
+        self.assertTrue(os.path.isdir(self.get_output('web/js')))
+        self.assertTrue(os.path.isfile(self.get_output('web/js/glue_data.js')))
+        self.assertTrue(os.path.isfile(self.get_output('web/index.html')))
+
     # Helpers
 
     def get_file(self, name='files'):
         return os.path.join(os.path.dirname(__file__), name)
+    
+    def get_output(self, filename):
+        return os.path.join(self.get_file('output'), filename)
 
-    def clean_output(self):
+    def generate(self, glue):
         directory = self.get_file('output')
-        shutil.rmtree(directory)
-        os.makedirs(directory)
-        return output
+        if os.path.isdir(directory):
+            shutil.rmtree(directory)
+        glue_directory = os.path.join(os.path.dirname(__file__), '..', '..')
+        output_directory = os.path.join(directory, 'glue')
+        web_directory = os.path.join(directory, 'web')
+        os.makedirs(output_directory)
+        glue.generate_files(glue_directory, output_directory, web_directory)
 
 if __name__ == '__main__':
     unittest.main()
